refactor: consolidate guest data transfer logic into utility
- Move guest data transfer from hooks/components to utils/guestDataTransfer.ts
- Update auth flows to use new centralized guest data transfer utility
- Remove GuestDataTransferHandler component and useGuestDataTransfer hook
- Simplify guest mode implementation across sign-in and verification flows
