+++ apps/expo/src/app/(auth)/verify-email.tsx
@@ -2,16 +2,19 @@ import React from "react";
 import { Pressable, Text, TextInput, View } from "react-native";
 import { KeyboardAwareScrollView } from "react-native-keyboard-controller";
 import { Redirect, Stack } from "expo-router";
-import { useSignUp } from "@clerk/clerk-expo";
+import { Clerk, useSignUp } from "@clerk/clerk-expo";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useConvexAuth } from "convex/react";
+import { useConvexAuth, useMutation } from "convex/react";
 import { usePostHog } from "posthog-react-native";
 import { Controller, useForm } from "react-hook-form";
 import { z } from "zod";
 
+import { api } from "@soonlist/backend/convex/_generated/api";
+
 import { useAppStore } from "~/store";
 import { Logo } from "../../components/Logo";
 import { logError } from "../../utils/errorLogging";
+import { transferGuestData } from "../../utils/guestDataTransfer";
 
 const verifyEmailSchema = z.object({
   code: z
@@ -30,6 +33,9 @@ const VerifyEmail = () => {
   const { isAuthenticated } = useConvexAuth();
   const hasCompletedOnboarding = useAppStore(
     (state) => state.hasCompletedOnboarding,
+  );
+  const transferGuestOnboardingData = useMutation(
+    api.guestOnboarding.transferGuestOnboardingData,
   );
   const {
     control,
@@ -63,6 +69,19 @@ const VerifyEmail = () => {
           email: completeSignUp.emailAddress,
           username: completeSignUp.username,
         });
+
+        // Transfer guest data after successful sign up
+        // Wait a bit for the session to be fully initialized
+        await new Promise((resolve) => setTimeout(resolve, 100));
+
+        // Safely access session data
+        const session = Clerk.session;
+        if (session?.user?.id) {
+          await transferGuestData({
+            userId: session.user.id,
+            transferGuestOnboardingData,
+          });
+        }
       } else {
         logError("Verification failed", completeSignUp);
         setGeneralError("Verification failed. Please try again.");
